feat: faas plugin refactoring with url path forwarding (#5616)
